app | 예약 조회 수정 & 플랜생성 알림제거

+++ src/services/ticketService.ts
@@ -67,6 +67,8 @@ async function update(options: {
   const connection = await db.beginTransaction()
   try {
     const {id, type, userId, trainerIds, franchiseId, totalSession, startedAt, expiredAt} = options
+    const reservationValidCount = await Reservation.findValidCount(id)
+    if (totalSession < reservationValidCount) throw new Error('not_allowed')
     await Ticket.update({id, type, totalSession, startedAt, expiredAt}, connection)
     await Ticket.deleteRelations(id, connection)
     await Ticket.createRelationExercises({userId, trainerIds, ticketId: id, franchiseId}, connection)
